requirements.txt
pandas 
matplotlib.pyplot 
seaborn 
streamlit 